% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawDice.R
\name{DrawDice}
\alias{DrawDice}
\title{DrawDice}
\usage{
DrawDice(n, pointsize = 0.5)
}
\arguments{
\item{n}{integer from 1-6}

\item{pointsize}{override for the size of the points}
}
\value{
plot
}
\description{
Draws a dice face given an integer from 1 to 6.
}
\examples{

DrawDice(5)
box()

}
