% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiceMap.R
\name{PlotDiceMap}
\alias{PlotDiceMap}
\title{PlotDiceMap}
\usage{
PlotDiceMap(image, pointsize = 0.5)
}
\arguments{
\item{image}{gsimage object}

\item{pointsize}{pointsize override}
}
\value{
plot
}
\description{
Takes a gsimage matrix image bitmap image object and returns a dice map plot.
}
\examples{

file = system.file("pictures", "pngexample.png", package="PictureToDice")
image = ImportPicture(file)
image.conv = ConvolutePixels(image,20)
image.prepped = PrepareDiceMap(image.conv, exposure.increase = 1)
PlotDiceMap(image.prepped, 0.2)


}
