% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotImage.R
\name{PlotImage}
\alias{PlotImage}
\title{PlotImage}
\usage{
PlotImage(image)
}
\arguments{
\item{image}{gsimage}
}
\value{
plot
}
\description{
Plots a grey scale image (gsimage)
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToDice"))
image.conv = ConvolutePixels(image,20)
PlotImage(image)

}
