% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDiceMap.R
\name{PrepareDiceMap}
\alias{PrepareDiceMap}
\title{PrepareDiceMap}
\usage{
PrepareDiceMap(image, exposure.increase = 0)
}
\arguments{
\item{image}{gsimage}

\item{exposure.increase}{Increase/Decrease the values of the dice}
}
\value{
gsimage.dice
}
\description{
PrepareDiceMap
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToDice"))
image.dice = PrepareDiceMap(image)
table(image.dice)

}
