% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToHexGrey.R
\name{ToHexGrey}
\alias{ToHexGrey}
\title{ToHexGrey}
\usage{
ToHexGrey(x)
}
\arguments{
\item{x}{numeric between 0 and 1}
}
\value{
character hexcode
}
\description{
Take a number x on the scale of 0-1 (black-white) and convert to hexcode
}
\examples{

# Black
ToHexGrey(0)
# White
ToHexGrey(1)

}
